backend/requirements.txt
fastapi>=0.111.0
# langchain>=0.2.3
# langchain_community>=0.2.4
# langchain_core>=0.2.5
# langchain_openai>=0.1.8
# langchain_pinecone>=0.1.1
# pinecone>=4.0.0
# pydantic>=2.7.4
python-dotenv>=1.0.1
# sentence-transformers
uvicorn